Improve crawl performance by reducing memory usage
