+++ src/main/java/com/scaleunlimited/flinkcrawler/fetcher/commoncrawl/SecondaryIndex.java
@@ -3,10 +3,18 @@ package com.scaleunlimited.flinkcrawler.fetcher.commoncrawl;
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public class SecondaryIndex {
 	
-	private String _indexFilename;
+    private static final Pattern FILENAME_PATTERN = Pattern.compile("cdx-(\\d{5}).gz");
+    
+    // WARNING - if anything changes in the serialization of this class, you must
+    // (a) bump the version number in SecondaryIndexMap, and (b) decide whether
+    // you'll support deserializing the older format.
+    
+    private int _indexFilenumber;
 	
 	private long _segmentOffset;
 	private long _segmentLength;
@@ -17,14 +25,19 @@ public class SecondaryIndex {
 	}
 	
 	public SecondaryIndex(String indexFilename, long segmentOffset, long segmentLength, int segmentId) {
-		_indexFilename = indexFilename;
+	    Matcher m = FILENAME_PATTERN.matcher(indexFilename);
+	    if (!m.matches()) {
+	        throw new IllegalArgumentException("Invalid index filename: " + indexFilename);
+	    }
+	    
+	    _indexFilenumber = Integer.parseInt(m.group(1));
 		_segmentOffset = segmentOffset;
 		_segmentLength = segmentLength;
 		_segmentId = segmentId;
 	}
 
 	public String getIndexFilename() {
-		return _indexFilename;
+		return String.format("cdx-%05d.gz", _indexFilenumber);
 	}
 
 	public long getSegmentOffset() {
@@ -40,14 +53,14 @@ public class SecondaryIndex {
 	}
 	
 	public void write(DataOutput out) throws IOException {
-		out.writeUTF(_indexFilename);
+		out.writeInt(_indexFilenumber);
 		out.writeLong(_segmentOffset);
 		out.writeLong(_segmentLength);
 		out.writeInt(_segmentId);
 	}
 	
 	public void read(DataInput in) throws IOException {
-		_indexFilename = in.readUTF();
+	    _indexFilenumber = in.readInt();
 		_segmentOffset = in.readLong();
 		_segmentLength = in.readLong();
 		_segmentId = in.readInt();
